mbid_mapping/requirements.txt
click==7.1.2
pytest==5.4.3
pytest-cov==2.10.0
psycopg2-binary==2.9.3
ujson==5.4.0
typesense
unidecode
python-dateutil==2.8.2
git+https://github.com/metabrainz/brainzutils-python.git@v2.1.0
tqdm==4.66.3
pyxDamerauLevenshtein==1.8.0
